refactor(desktop): extract runYtDlpDownload service from downloader handler
